+++ lib/app.dart
@@ -252,7 +252,7 @@ class _WalletAppState extends State<WalletApp> {
         ReferendumVotePage.route: (_) => ReferendumVotePage(_appStore),
         // profile
         AccountManagePage.route: (_) => AccountManagePage(_appStore),
-        ContactsPage.route: (_) => ContactsPage(_appStore.settings),
+        ContactsPage.route: (_) => ContactsPage(_appStore),
         ContactListPage.route: (_) => ContactListPage(_appStore),
         ContactPage.route: (_) => ContactPage(_appStore),
         ChangeNamePage.route: (_) => ChangeNamePage(_appStore.account),