+++ lib/page/profile/recovery/recoveryStatePage.dart
@@ -81,15 +81,15 @@ class _RecoveryStatePage extends State<RecoveryStatePage> {
           .toList();
       List statusList = List.of(status[2]);
 
+      int invalidCount = 0;
       statusList.toList().asMap().forEach((k, v) {
         // recovery status is null if recovery was closed
         if (v == null) {
           print('remove $k');
-          ls.removeAt(k);
-          if (infoList.length > k) {
-            infoList.removeAt(k);
-          }
-          statusList.removeAt(k);
+          ls.removeAt(k - invalidCount);
+          infoList.removeAt(k - invalidCount);
+          statusList.removeAt(k - invalidCount);
+          invalidCount++;
         }
       });