+++ lib/utils/i18n/home.dart
@@ -46,6 +46,7 @@ const Map<String, String> enHome = {
   'update.error': 'Update Failed',
   'exit.confirm': 'Do you want to exit the App?',
   'input.invalid': 'Invalid input',
+  'loading': 'Loading...',
 };
 
 const Map<String, String> zhHome = {
@@ -93,4 +94,5 @@ const Map<String, String> zhHome = {
   'update.error': '更新失败',
   'exit.confirm': '确定要退出 App 吗？',
   'input.invalid': '输入格式错误',
+  'loading': '载入中。。。',
 };